Add simple validation on fields

+++ front/src/app/modules/customer/components/personal-info/users/users.component.ts
@@ -2,6 +2,7 @@ import {Component, EventEmitter, Input, OnDestroy, OnInit, Output} from '@angula
 import {UsersService} from '../../../../../services/users.service';
 import {User} from '../../../models/user';
 import {Subscription} from 'rxjs';
+import {StorageService} from '../../../../../services/storage.service';
 
 
 @Component ({
@@ -15,7 +16,8 @@ export class UsersComponent implements OnInit {
   public users: User[];
   @Output() public showUsersEvent: EventEmitter<any> = new EventEmitter<any>();
 
-  constructor(private usersService: UsersService) {}
+  constructor(private usersService: UsersService,
+              private storageService: StorageService) {}
 
   ngOnInit(): void {
     this.loadUsers();
@@ -29,7 +31,7 @@ export class UsersComponent implements OnInit {
 
   public deleteUsers(): void {
     for (const user of this.users) {
-      if (user.isDelete) {
+      if (user.isDelete && user.login !== this.storageService.getCurrentUser().login) {
         this.usersService.deleteUser(user.id).subscribe(res => {
           this.loadUsers();
         });
